(1) Refactor, change the badge rotation speed (2) Refactor, display canvas dispose and buffer setting handle (3) Refactor, move the iconTexture into Utility folder
